Enabled better time step division; rate and bhp targets simulator-side.
In WellControls changed int to double which allows for division of two, independent of even or odd days interval. Previously the timestep controls was set to 2*[number] which put all max timestep to 2 days. By setting [number]*2 it divides the interval by two and sets half of that time as max timestep.

Previously it was only possible to specify an argument for the control mode that was chosen. Instead, this allows for setting an upper limit independent of the control mode (either BHP or Rate). This enables better optimization when the constraints are given in BOTH rate and BHP. This enables setting rate/BHP limits simulator side instead of using FieldOpt constraints.

+++ FieldOpt/Simulation/simulator_interfaces/driver_file_writers/driver_parts/ecl_driver_parts/wellcontrols.h
@@ -64,7 +64,7 @@ namespace Simulation {
                          * \param time The time step to be inserted.
                          * \return
                          */
-                        virtual QString createTimeEntry(int time, int prev_time);
+                        virtual QString createTimeEntry(double time, double prev_time);
                         QString createProducerEntry(WellSetting *setting);
 
                         /*!
